smallarms: port
