+++ addons/smallarms/$PBOPREFIX$
@@ -1,0 +1,1 @@
+z\ace\addons\switchunits
