+++ addons/smallarms/CfgVehicles.hpp
@@ -1,0 +1,34 @@
+/*#define MACRO_ADDITEM(ITEM,COUNT) class _xx_##ITEM { \
+    name = #ITEM; \
+    count = COUNT; \
+};
+
+class CfgVehicles {
+    class NATO_Box_Base;
+    class EAST_Box_Base;
+    class IND_Box_Base;
+
+    class Box_NATO_Support_F: NATO_Box_Base {
+        class TransportItems {
+            MACRO_ADDITEM(ACE_acc_flashlight_tls,5)
+        };
+    };
+
+    class Box_East_Support_F: EAST_Box_Base {
+        class TransportItems {
+            MACRO_ADDITEM(ACE_acc_flashlight_tls,5)
+        };
+    };
+
+    class Box_IND_Support_F: IND_Box_Base {
+        class TransportItems {
+            MACRO_ADDITEM(ACE_acc_flashlight_tls,5)
+        };
+    };
+
+    class ACE_Box_Misc: Box_NATO_Support_F {
+        class TransportItems {
+            MACRO_ADDITEM(ACE_acc_flashlight_tls,2)
+        };
+    };
+};*/
