+++ addons/smallarms/script_component.hpp
@@ -1,0 +1,12 @@
+#define COMPONENT smallarms
+#include "\z\ace\addons\main\script_mod.hpp"
+
+#ifdef DEBUG_ENABLED_SMALLARMS
+	#define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_SMALLARMS
+	#define DEBUG_SETTINGS DEBUG_SETTINGS_SMALLARMS
+#endif
+
+#include "\z\ace\addons\main\script_macros.hpp"
